refactor(scripts): split init script
